Fixed some issues with experience modifiers that can no longer be ignored now that the Maximizer deals with familiar items:
The "Experience" modifier now applies only to sources of substat points that
are subject to the normal 2:1:1 stat distribution.  Three new numeric
modifiers ("Muscle Experience", etc.) have been added for sources of specific
substats.  One of these new modifiers, corresponding to your mainstat, is
what mafia now considers to be your EXP bonus - as shown in the sidepane and
in Location Details, as returned by experience_bonus(), and as boosted by
"maximize exp".  This value will be roughly half of what it was before, but
now it accurately represents the experience that actually counts towards
gaining levels.

Added two new string modifiers to describe nonstandard stat distributions:
"Stat Tuning" (for the Crimbo '06 offhands that give you a different class's
distribution) and "Familiar Tuning" (for the Bandersnatch's jub-jub birds).

The Maximizer no longer considers leaving a slot empty if it has anything at
all that appears useful for that slot.  This greatly reduces the number of
combinations it has to check, but could conceivably cause problems in-run if
you have only poor choices for a given slot.  Please report any anomalies.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@7768 b29ace70-8910-0410-8dcc-aa2fc6433167