Added some of today's buggy items, although not the familiar itself.
Added ConsequenceManager, with data file consequences.txt, as a framework for
applying a regular expression to specific pages and performing actions based
on the results, such as setting preferences.  The currently supported
consequences can trigger on an item, skill, or effect description page; the
skill level check for the slimy skills is now in this form, and there are
also now checks for the chess cookie effects, ballpit bonus, and the
Landscaper's leafblower +ML value.

On the first login of each day, one of these descriptions will be loaded
(sequentially), so mafia will eventually be synchronized with the player
state even without the player opening the descriptions while mafia is
running.

Setting debugConsequences=true will print a message whenever a consequence is
triggered.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@8349 b29ace70-8910-0410-8dcc-aa2fc6433167