A CCS section can now contain many of the combat macro commands, which are passed thru unchanged in a macrofied combat, or ignored completely if the combat can't be macrofied: scrollwhendone, mark, goto, call, sub/endsub, if/endif, and while/endwhile.  There are two intentional omissions: "icon" (which is meaningless in an on-the-fly macro) and "repeat" (which isn't safe to use with mafia's combat actions, as they may expand to more than one macro action).  If you want an action to be repeated until the combat is over, either put an explicit loop around it, or simply let it be the final command in the CCS section in which case mafia will loop it for you.
An arbitrary macro command can be inserted in a CCS by enclosing it in double
quotes.  This would be useful for newly-added commands (prior to explicit
mafia support being added), and also to override any mafia augmentation of
the command.  For example, "pickpocket" (with the quotes) would
unconditionally attempt to steal an item, even if mafia thinks the monster
has no drops, or that all drops are guaranteed due to your item drop
modifier.  (You might be planning to run away after a successful pickpocket,
perhaps.)

If you have a CCS section named [global prefix], its contents will be
macrofied prior to the section that matches the monster name or location (or
ignored in round-by-round combat).  You can use it to define subroutines that
can be used from any CCS section.  It would also be the ideal place to put a
"scrollwhendone" command if you want that globally enabled.

The CCS editor no longer inserts line numbers in each section, as they are no
longer very useful in the presence of macro loops, conditionals, and
subroutines that make the line number entirely unrelated to the round number.
Existing line numbers will disappear the first time you edit and save the
CCS.  You can still manually type a number followed by a colon at the start
of a line, as a shortcut for replicating the previous line until the
specified line number is reached.

The CCS editor's "help" button now works in both view and edit modes.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@8452 b29ace70-8910-0410-8dcc-aa2fc6433167