Add a display of your current tokens and tickets in the arcade.
When starting a game of DemonStar, give a warning if the player has a weapon
with a power not between 50 and 150, or has no hat or pants equipped.  The
details are subject to change as the game becomes better understood, but this
should prevent wasting turns and a token on a game that would be impossible
to win.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@8478 b29ace70-8910-0410-8dcc-aa2fc6433167