Since it's no longer a property in KoLmafia, expose the current auto attack settings as get_auto_attack() and set_auto_attack() so that users can check/reset the auto attack (if they want to have the old entangling noodles behavior, for example).
While it returns a string, I believe that most (if not all) of the time it's an integer value.  If it is disabled, it will actually return "0" and if it is set to attack, it will actually return "1".  For non-macros, it will return the skillId as a string, and I'm not sure what it returns for macros.

If it turns out that it's always an integer value, this method signature will be changed to always return an integer.  Just let me know on the kolmafia.us forums.

git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@8970 b29ace70-8910-0410-8dcc-aa2fc6433167