Emergency account option fix:
On login, look at the Interface, Inventory, Combat, and Account tabs. 
- The Interface tab tells what type of top menu you have and whether you have
  a compact charpane
- The Inventory tab says whether you have compact autosell and whether unequip
  all also strips your familiar
- The Combat tab tells you whether you have the CAB enabled and what your 
  autoattack is
- I think the Account menu tells you if you are in Bad Moon, whether you have
  recalled skills, whether you are in Hardcore, and if you have any consimption
  restrictions.

If you manually visit the options, reparse whichever tab you visit.

We do not yet recognize when you change something on one of the tabs, but if you
move to a different tab and back again, we will see the change.

The Bad Moon, Hardcore, and Consumption Restriction recognition almost certainly
does not work; I have no data about whether those things are really on the 
Account Option tab or another one, or what the HTML looks like for the tab when
any of those features are active.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@9001 b29ace70-8910-0410-8dcc-aa2fc6433167