Give a better estimate of familiar experience gained on a combat win using modifiers rather than a hard-coded list.
Handle 100% familiar run checking with a single value in preferences (singleFamiliarRun) which represents the current familiar being used during the run.

If the familiar is not yet known, this value is set to zero so it can be detected as needed.  Set this value to -1 when more than one familiar is detected as being used for something that initiates a combat, though this recognizes items that initiate combats (which I'm not sure count, but if you're using a different familiar for that combat, you probably didn't care too much).

Since it's down to a single value, you could theoretically cheat KoLmafia by setting the value to -1 (which is the same value that KoLmafia will set the preference to if it detects you start using more than one familiar), making this work kinda like an invisible preference, but since it resets every run, it's rather useless as one.

git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@9779 b29ace70-8910-0410-8dcc-aa2fc6433167