The create vs. buy decision now applies in all cases where both creation and Mall purchase are valid options, not merely the cases where you have some (but not all) of the needed ingredients.  The buyScript mechanism is still supported, but should hopefully be less needed now.
The decision to buy a completed item rather than creating it from ingredients
already in inventory requires assigning a value to those ingredients, which
really depends on play style.  Not everyone is going to put in the effort
needed to maximize their Mall profits; they might use only autosell to
dispose of excess items, or just hoard them.  Therefore, a new float
preference "valueOfInventory" allows players to indicate the worth of items,
with these key values:
0.0 - Items already in inventory are considered free.
1.0 - Items are valued at their autosell price.
2.0 - Items are valued at current Mall price, unless they are min-priced.
3.0 - Items are always valued at Mall price (not really realistic).
Intermediate values interpolate between integral values.  The default is 1.8,
reflecting the fact that items won't sell immediately in the Mall without
undercutting or advertising.  This preference, and several previously hidden
prefs affecting create vs. buy decisions, are now exposed on a new Creatable
-> Fine Tuning page in the Item Manager.  (The page will eventually have a
list of manual creations, which is why I put it there instead of
Preferences.)

To offset the increased frequency of create vs. buy decisions, they now make
greater use of historical prices.  If there is a greater than 2:1 ratio in
the cost of the two options, using prices up to 7 days old, the decision will
be made without hitting the server with a Mall search.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@9806 b29ace70-8910-0410-8dcc-aa2fc6433167