Since CLI commands, buttons, menu items and action panels all run in something other than the event dispatch thread, remove FoxTrot library.
If the UI locks up when you do something because it was overlooked in all this refactoring, that would be a bug.

If the UI renders funny because things aren't running in the event dispatch any more, that might also be a bug.

If you get a stack trace that says something like "pause object in event dispatch thread", that might also be (you guessed it) a bug.

git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@9944 b29ace70-8910-0410-8dcc-aa2fc6433167