When we log in, we can get redirected to fight.php or choice.php. When that happens, we abort refreshing the character until after the fight or choice is resolved in the Relay Browser. However, after the aborted character refresh, KoLmafia proceeds to build the various frames that the user wants, either as frames or as stabdalone frames. Some of those frames need to make requests to construct controls - such as the Skill Casting frame, which wants to read your account contacts in order to make the dropdown of victims - which will, themselves, get redirected if they are attempted while in a fight or a choice.
Create a mechanism for a request to get posted after initialization is complete.
Make ContactListRequest use this mechanism, so that the Skill Casting frame's
victim list will get updated after we finish logging in, even if the frame was
constructed durin a redirected login.

There is plenty more to come, along these lines...


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@10324 b29ace70-8910-0410-8dcc-aa2fc6433167