Certain frames - Coinmasters, Gear Changer, Daily Deeds, at least - depend on information available only after you have logged in and refreshed your session. If you log in and are redirected to a fight or choice, we short circuit the session refresh until after the fight or choice is complete. After the short- circuited login, if you happen to have one of the above mentioned frames set to load as a tab or independant frame, they will get constructed - still missing essential information. Create a new mechanism:
RequestThread.executeMethodAfterInitialization( object, methodName )

which will execute a method (normal or static) after you are logged and your
session is refreshed. Make the above mentioned frames use this mechanism.

Add a "guild" link to your pants when you steal them from yourself to fulfil
the guild challenge.

Do not let the LogoutRequest force KoLmafia into an ABORT state. Not only does
that prevent a script from executing further - hopefully, a login command - but
it suppresses several of the messages in the Logout Manager, including the one
which tells you that the Logout process is complete.

Undo a change I made to AdventureRequest in 9418 whch apparently broke adv1.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@10331 b29ace70-8910-0410-8dcc-aa2fc6433167