Using the HTML cleaner on fight pages is no longer experimental. Therefore, the "option" to clean and log it is really just the option to log the tree. Correct the description on the Options Frame
The various "use" CLI commands throw an exception if given no arguments. Fix.

When you find a pasta guardian item, give a "summon" use link.

When we are unable to find a standard fight tree in the fight.php HTML, call
the default result processing method, rather than depending on the fight parser
(which won't be called) to do it.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@10366 b29ace70-8910-0410-8dcc-aa2fc6433167