gnasirProgress is a bitmap of things you've done with Gnasir to advance desert exploration:
stone rose = 1
black paint = 2
killing jar = 4
worm-riding manual pages (15) = 8
successful wormride = 16

When you find a stone rose, add [gnasir] link to visit him
When you find manual pages
- If you have less than 15, add informational [x] to tell you how many you have
- If you have 15, add [gnasir] link to visit him
When you are given the worm-riding hooks and the desert is less than 100% explored:
- If you have a drum machine, "wormride" link uses a drum machine
- Otherwise "oasis" link takes you to Oasis to find one
When you find a drum machine and the desert is 100% explored
- "use" link will use it and let you fight a giant sandworm
When you find a drum machine and the desert is less than 100% explored
- If you have worm riding hooks, "wormride" link uses a drum machine
- Otherwise, no link (lets you save it for worm riding)
When you get a desert sightseeing pamphlet, give it a use link.
When you finish your desert exploration, you get a New Area Unlocked message for A Small Pyramid. Linkify that to go to the pyramid
When you click on the image/link to visit the small pyramid, if you have "autoCraft" enabled, make a Staff of Ed first, if you can.


git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@13303 b29ace70-8910-0410-8dcc-aa2fc6433167