Change Maximizer handling of Sticker Weapons, Card Sleeve and Folder Holder. Previously they were always passed as possible items to the speculation loop if any of the modifiers of ANY of the stickers/cards/folders were useful to the maximization. Now they are only passed if the modifiers of the CURRENT stickers/cards/folders are good enough to be worth using. Should reduce the number of combinations used a bit.
Note - There is an argument for treating card sleeves in a similar way to the new Throne/Bjorn code, so swapping cards could be considered. It's a good argument, and wouldn't add much time to maximize processing, but I don't have one to test with, so am not doing so.

git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@13655 b29ace70-8910-0410-8dcc-aa2fc6433167