Add multi-line string support to ASH.
In order to specify a multi-line string, you must explicitly escape the
end-of-line by placing a backslash (\) at the end of the line.

Note that as a consequence of how the parser is written, it will trim all
leading and trailing whitespace for each of these lines via Java's
String.trim() method. Previously, this was irrelevant in the context of ASH
strings, but now that ASH strings can span multiple lines, it is worth
mentioning. If you want to have whitespace at the beginning of the line,
just escape the first character of the line.

git-svn-id: svn://svn.code.sf.net/p/kolmafia/code@16800 b29ace70-8910-0410-8dcc-aa2fc6433167